refactored views
and changed the way the app is called in the init file - now the chunks
of app logic are called in a callback function, allowing the app to
initialize all the way before things like views and layers are
initialized.  I think we also need bools on each big chunks of app so we
can reset (or not) properly when a new file is loaded in the same
session.