embed example 001 working
with refactored code.  so is the viewer.  i experimented a little with
chunks of app logic in the embed viewer ... and started a css wild goose
chase.  have to do some improvements, but both sites are running from
the same javascript library!  feels like good progress, but more work
ahead.